+++ lua/plugins/lualine.lua
@@ -1,8 +1,10 @@
 -- lua/plugins/lualine.lua
 
-local use = require('packer').use
-
-use({
+return {
     'nvim-lualine/lualine.nvim',
-    requires = { 'nvim-tree/nvim-web-devicons', opt = true },
-})
+    lazy = false,
+    dependencies = { 'nvim-tree/nvim-web-devicons' },
+    config = function()
+        require('config.lualine')
+    end,
+}