+++ lua/plugins/luasnip.lua
@@ -1,14 +1,9 @@
 -- lua/plugins/luasnip.lua
 
-local use = require('packer').use
-
-use({
+return {
     'L3MON4D3/LuaSnip',
-    requires = {
-        'rafamadriz/friendly-snippets',
-    },
-    config = function ()
+    dependencies = { 'rafamadriz/friendly-snippets' },
+    config = function()
         require('luasnip/loaders/from_vscode').lazy_load()
-    end
-})
-
+    end,
+}