+++ lua/plugins/packer.lua
@@ -1,8 +1,0 @@
--- lua/plugins/packer.lua
-
-local use = require('packer').use
-
-use({
-    'wbthomason/packer.nvim',
-})
-