+++ lua/plugins/treesitter-textobjects.lua
@@ -1,11 +1,6 @@
 -- lua/plugins/treesitter-textobjects.lua
 
-local use = require('packer').use
-
-use({
+return {
     'nvim-treesitter/nvim-treesitter-textobjects',
-    requires = {
-        'nvim-treesitter/nvim-treesitter',
-    },
-})
-
+    dependencies = { 'nvim-treesitter/nvim-treesitter' },
+}